[INTERPRETER] * NEW: Remove as much "packed" structure declaration as possible, to   prevent possible bugs on ARM architecture.
git-svn-id: svn://localhost/gambas/trunk@5860 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_file_temp.h
@@ -519,8 +519,8 @@ void FILE_stat(const char *path, FILE_STAT *info, bool follow)
 	info->mtime = (int)buf.st_mtime;
 	info->ctime = (int)buf.st_ctime;
 	info->hidden = (*FILE_get_name(path) == '.');
-	info->uid = (int)buf.st_uid;
-	info->gid = (int)buf.st_gid;
+	info->uid = buf.st_uid;
+	info->gid = buf.st_gid;
 }
 
 char *FILE_mode_to_string(mode_t mode)
